backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
PyJWT>=2.8.0
bcrypt>=4.1.0
python-dotenv>=1.0.0
